archive/sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paw-sdk"
version = "1.0.0"
description = "Official Python SDK for PAW blockchain"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["paw", "blockchain", "cosmos", "sdk", "defi", "dex"]
authors = [
    {name = "PAW Chain", email = "dev@paw.network"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "httpx>=0.24.0",
    "bech32>=1.2.0",
    "ecdsa>=0.18.0",
    "mnemonic>=0.20",
    "protobuf>=4.23.0",
    "cosmospy-protobuf>=1.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "mypy>=1.4.0",
    "pylint>=2.17.0",
]

[project.urls]
Homepage = "https://github.com/paw-chain/paw"
Documentation = "https://docs.paw.network"
Repository = "https://github.com/paw-chain/paw"
Issues = "https://github.com/paw-chain/paw/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["paw"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
